feat: incluindo esqueleto
